⌛Refactor: DailyCheck 에 사용자 식별 로직 추가

+++ src/main/kotlin/com/example/days/domain/dailycheck/service/DailyCheckServiceImpl.kt
@@ -16,8 +16,7 @@ class DailyCheckServiceImpl(
     private val resolutionRepository: ResolutionRepository
 ):DailyCheckService {
     @Transactional
-    override fun createDailyCheck(resolutionId: Long, request: DailyCheckRequest): DailyCheckResponse {
-        val userId = getAuthenticationUserId()
+    override fun createDailyCheck(resolutionId: Long, userId: Long, request: DailyCheckRequest): DailyCheckResponse {
         val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
 
         if(userId == resolution.author.id){
@@ -29,22 +28,36 @@ class DailyCheckServiceImpl(
             ?: TODO("예외처리")
     }
 
-    override fun getDailyCheckByList(resolutionId: Long): List<DailyCheckResponse> {
-        return dailyCheckRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"))
-            .map { DailyCheckResponse.from(it) }
+    override fun getDailyCheckByList(resolutionId: Long, userId: Long): List<DailyCheckResponse> {
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
+        if(userId == resolution.author.id){
+            return dailyCheckRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"))
+                .map { DailyCheckResponse.from(it) }
+        }
+        else TODO()
     }
 
     @Transactional
-    override fun updateDailyCheck(resolutionId: Long, dailyCheckId: Long, request: DailyCheckRequest)
+    override fun updateDailyCheck(resolutionId: Long, userId: Long, dailyCheckId: Long, request: DailyCheckRequest)
     : DailyCheckResponse {
-        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO()
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
         val dailyCheck = dailyCheckRepository.findByIdOrNull(dailyCheckId) ?: TODO()
-        dailyCheck.updateDailyCheck(request.memo, resolution)
-        return DailyCheckResponse.from(dailyCheck)
+        if(userId == resolution.author.id){
+            dailyCheck.updateDailyCheck(request.memo, resolution)
+            return DailyCheckResponse.from(dailyCheck)
+        }
+        else TODO()
+
+
     }
 
     @Transactional
-    override fun deleteDailyCheck(dailyCheckId: Long) {
-        dailyCheckRepository.deleteById(dailyCheckId)
+    override fun deleteDailyCheck(resolutionId: Long, dailyCheckId: Long, userId: Long) {
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
+
+        if(userId == resolution.author.id) {
+            dailyCheckRepository.deleteById(dailyCheckId)
+        }
+        else TODO()
     }
 }
